Update Clang's handling of C99 6.5.15p6 to support memory capabilities:
If both operands are pointers to compatible types or to differently qualified
versions of compatible types, the result type is a pointer to an appropriately
qualified version of the composite type.

We update this to also qualify the pointer type with the memory_capability
attribute.

Added test for when ternary operands are capabilities.