Add ability to represent and qualify C++ references as capabilities.
This addresses two things in C++:

1) Reference types were being already being lowered to capabilities in LLVM IR
when in the Pure Capability ABI, but some Clang optimisations require knowing
that a reference is a capability during semantic analysis, otherwise it can
lead to incorrectly generated IR.

2) There was no way to express in the Hybrid ABI that a reference type should
be lowered to a capability.

Fixing both of these requires explicitly representing that a reference type is
a capability. This has been implemented in this commit as a bitfield in the
ReferenceType class (superclass of LValueReferenceType and RValueReferenceType)
that gets set if in the pure capability ABI or if the reference type is
__capability-qualified (in the Hybrid ABI).

The __capability qualifier is written just like with pointers. For example:

void f(A& __capability a) { ... }

Also added pretty printing of a __capability-qualified reference type when the
AST is dumped.