Revert "Prevent clang from coercing a C++ class or struct argument that contains"
This commit does not correctly fix the issue of capabilities in superclasses
being turned into { i64, i64, i64, i64 } (when objects are passed by value).
The real fix is to make clang have the same behaviour as in the single-level
class hierarchy case (i.e. capabilities are turned into i8 addrspace(200)*).
This ensures that registers are used if the class/struct is small enough and
otherwise they are passed indirectly. This would also be consistent with C
structs.

This reverts commit 87215a258f345ebedeee0d99b1026d30a7b1488b.