Still support __cheri_cast but output a deprecated warning with a Fix-It hint.
The deprecated looks like the following:

file.c:6:14: warning: __cheri_cast is deprecated; use __cheri_ptr instead [-Wdeprecated]
  char *d = (__cheri_cast char *)c;
             ^~~~~~~~~~~~
             __cheri_ptr

The Fix-It hint not only shows what to fix but it can also be applied to the
source file by passing -Xclang -fixit to clang.  This will replace __cheri_cast
with __cheri_ptr in-place.