Handle the new alignment builtins in integer constant expressions
C code should be able to use these aligment builtins as a replacement for
macros that do bitwise arithmetic. If we don't evalute them as constants
it is not possible to initialize global variables with the builtins.

Now it is possible to write the following C code:

`int x  = __builtin_align_up(129, 128) // x will be set to 256`