Implement the alignment builtins for CHERI capabilities
They are currently still poorly optmized for __uintcap_t when used on
constant integers but that can be fixed in FoldCapIntrinsics.cpp.
When creating the newly aligned capability in __builtin_align_up/down I am
not sure if it is better to use llvm.cheri.cap.offset.increment or an
inbounds GEP since both have cases that are not optimized properly and
result in bad code. For now I decided to use inbounds GEP since more LLVM
passes are likely to understand it.