Introduce an experimental virtual address instead of offset compilation mode
This mode can be toggled using the -cheri-uintcap=addr (new) and
-cheri-uintcap=offset (old, default value) flags.

In this mode all operations on uintcap_t will be performed on the virtual
address instead of the offset. This should fix problems such as alignemnt
checks using uintcap_t and using xor for pointer hashing. The downside is
that we are now more slightly more likely to leak virtual addresses, which
could make copying garbage collection more difficult. However, we there are
already lots of cases where this happens (e.g. hash tables that use the
pointer value) so it might make sense to make this the default compilation
mode at some point in order to increase C compatibility.