Use libcalls for atomic RMW operations on capabilities
Otherwise we assert during codegen. Silence the inefficient atomics warning
in this case since it's a known issue and breaks the -Werror build.

Also extend the test to check the assembly code generation for atomics.
The atomic load/store work for capabilities. Also atomic rmw on i64/i32 via
a capability pointer also work with clld/cscd, it's only the cllc/cssc
that is broken and will be non-trivial to fix.