Change default uintcap_t mode to address instead of offset
Using offset has resulted in tons of annoying to debug failures in the past
and gives highly surprising results for casts from void* -> intptr_t.
For example, the result of (vaddr_t)(intptr_t)(void*) is the offset of the
capability (often 0) while the user almost certainly wanted the virtual
address as the result of this cast.
This is just one of the many failures we have encounted in the past that
could have been avoided by using an address interpretation.
Offset interpretation is still useful if you would like to support copying
GC in C since it avoids direct use of virtual addresses.

The old behaviour can still be turn on by passing -cheri-uintcap=offset
to the compiler and will not be removed. However, the cost of using the
offset mode by default far outweighs the potential advantages (we are not
using copying GC in C anyway) and using the address is much closer to the
expectations of most C programmers.