Return Optional<int64_t> from canTightenCheriBounds()
This will allow setting the bounds to the container type instead of just
individual member or not at all. We need this for union types where we do
want to shrink the bounds to the union but using the individual member
may be unsafe (e.g. struct sockaddr). It might also be useful for arrays so
that we can set bounds to the whole array if it is embedded in a struct.